python/requirements.txt
holidays
spyne
twisted
numpy
ssdb
python-dateutil
protobuf==3.0.0

